requirements.txt
discord.py
langchain
langchain-community
ollama
pdfminer.six
pydantic
pypdf
gpt4all
chromadb